backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
django-cloudinary-storage>=0.3.0
cloudinary>=1.36.0
Pillow>=10.0.0
python-decouple>=3.8
djangorestframework-simplejwt>=5.3.1
holidays>=0.34